[#19]Refactor:Enum 정리, 회원-회원상세 연관관계 수정
