+++ src/main/java/com/server/money_touch/domain/user/enums/Role.java
@@ -1,0 +1,5 @@
+package com.server.money_touch.domain.user.enums;
+
+public enum Role {
+    ADMIN, USER;
+}
